1.7 beta release changes & added features.
it's the 1.7 version of this program. I chose 1.7 simply because
VibranceGUI is close to be finished and I want 2.0 to be the first
stable version.

You can now affect only a specific monitor, or all in once.
few changes in the GUI, and other stability improvements.
Now working on specific application vibrance settings, e.g. apply
vibrance only when CS2 is full-screen and running.

Signed-off-by: Roi L <roiinpushq@proton.me>