beta-1.8 release changes & notes
I've added the ability to enter a desired Process ID inside
a GtkEntry widget, you can now hook applications and activate
vibrance if the window is focused and is in full-screen mode, this is
pretty cool.

- I understand there may be memory leaks and inefficient code I'm not aware of and I'm
trying to look up into it. Note that currently there's a lot of
debugging going on.

Also note that the specific application configuration feature is not done yet; what left is
to allow a specific vibrance level to be set instead of the default
maximum level; and graphically show the user which applications are
currently under interception.

Signed-off-by: Roi L <roiinpushq@proton.me>