improvements & hash table functionality
Now you can use and add PID you want to track with the desired
Vibrance level, focus on the program, and boom.

For now the functionality for removing keys off the hash table
wasn't implemented yet, and the need for memory leaks analyize,
efficient code, and more features will be added soon.

Signed-off-by: Roi L <roiinpushq@proton.me>